(#10) refactor :: 커스텀 예외 패키지 구조 수정
