+++ src/main/java/com/example/sillok_server/infra/exception/ImageNotFoundException.java
@@ -1,7 +1,7 @@
 package com.example.sillok_server.infra.exception;
 
-import com.example.sillok_server.global.error.ErrorCode;
-import com.example.sillok_server.global.error.SillokException;
+import com.example.sillok_server.global.error.exception.ErrorCode;
+import com.example.sillok_server.global.error.exception.SillokException;
 
 public class ImageNotFoundException extends SillokException {
